+++ hadar/solver/actor/domain.py
@@ -18,34 +18,6 @@ class DTO:
         return self.__str__()
 
 
-class Consumption(DTO):
-
-    def __init__(self, quantity: int, cost: int=0, type: str=''):
-        self.cost = cost
-        self.quantity = quantity
-        self.type = type
-
-
-
-
-
-class Production(DTO):
-
-    def __init__(self, quantity: int, cost: int=0, type: str='in', id: uuid=0, exchange: Exchange = None):
-        self.type = type
-        self.cost = cost
-        self.quantity = quantity
-        self.id = id
-        self.exchange = exchange
-
-
-class Border(DTO):
-    def __init__(self, dest: str, capacity: int, cost: int=0):
-        self.dest = dest
-        self.capacity = capacity
-        self.cost = cost
-
-
 class NodeQuantity(DTO):
     def __init__(self, consumptions: List[Consumption]=[], productions: List[Production]=[], borders: [Border]=[], min_exchange=1):
         self.min_exchange = min_exchange
@@ -69,22 +41,3 @@ class NodeState(DTO):
         self.rac = rac
 
 
-class Event(DTO):
-    def __init__(self, type: str, message, res=None):
-        self.type = type
-        self.message = message
-        self.res = res
-
-
-class Snapshot(DTO):
-    def __init__(self):
-        pass
-
-
-class Start(DTO):
-    def __init__(self):
-        pass
-
-class Next(DTO):
-    def __init__(self):
-        pass
