+++ hadar/solver/actor/domain/message.py
