+++ tests/solver/actor/test_actor.py
@@ -1,0 +1,9 @@
+import numpy as np
+import unittest
+
+from hadar.solver.actor.domain.input import *
+
+class TestDispatcher(unittest.TestCase):
+
+    def test_build_state(self):
+        pass # TODO
